fix: Mapper, UseCase 를 Impl과 통합 및 레코드변환
