+++ src/main/java/leets/bookmark/application/dto/response/CategoryResponse.java
@@ -1,15 +1,4 @@
 package leets.bookmark.application.dto.response;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Builder;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class CategoryResponse {
-    private Long id;
-    private String categoryName;
+public record CategoryResponse(Long id, String categoryName) {
 }
