+++ src/main/java/leets/bookmark/application/usecase/CreateCategoryUseCase.java
@@ -1,8 +1,19 @@
 package leets.bookmark.application.usecase;
 
 import leets.bookmark.application.dto.request.CreateCategoryRequest;
-import leets.bookmark.application.dto.response.CategoryResponse;
+import leets.bookmark.application.mapper.CategoryMapper;
+import leets.bookmark.domain.service.CategorySaveService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface CreateCategoryUseCase {
-    void save(Long userId, CreateCategoryRequest request);
+@Service
+@RequiredArgsConstructor
+public class CreateCategoryUseCase {
+
+    private final CategorySaveService categorySaveService;
+    private final CategoryMapper categoryMapper;
+
+    public void save(Long userId, CreateCategoryRequest request) {
+        categorySaveService.save(categoryMapper.toCategory(userId, request));
+    }
 }
